+++ Semana 1/react/src/Components/Body.js
@@ -1,11 +1,21 @@
 import React from "react";
-import {getTime} from '../time'
+import { getTime } from "../time";
+import api from "../services/api";
 
 export default function Body(props) {
-  function handleAddProject(name) {
-      //props.projects.push(`${name} - ${Date.now()}`)//Nao respeita o conceito de imutabilidade
-        props.setProjects([...props.projects,`${name} - ${getTime()}`]); // Solicita ao react que renderize o componente novamente passando as informacoes adicionais
-     
+  async function handleAddProject(name) {
+    //props.projects.push(`${name} - ${Date.now()}`)//Nao respeita o conceito de imutabilidade
+
+    //props.setProjects([...props.projects,`${name} - ${getTime()}`]); // Solicita ao react que renderize o componente novamente passando as informacoes adicionais
+
+    const response = await api.post("projects", {
+      title: `Novo Projeto - ${getTime()}`,
+      owner: "Rafael Muniz",
+    });
+
+    const project = response.data
+
+    props.setProjects([...props.projects,project])
   }
 
   return (
@@ -13,9 +23,9 @@ export default function Body(props) {
       <h2>{props.title}</h2>
       <ul>
         {props.projects.map((project) => (
-          <li key={project}>{project}</li>
+          <li key={project.id}>{project.title}</li>
         ))}
-      </ul> 
+      </ul>
 
       <button type="button" onClick={handleAddProject}>
         Adicionar Projeto